Set up database models and helpers
