+++ src/server/model/action.js
@@ -1,0 +1,11 @@
+'use strict';
+
+import { actionQueryBuilder } from '../../db/helpers/actionQueryBuilder';
+
+module.exports = knex => {
+	const models = actionQueryBuilder(knex);
+
+	return {
+		...models
+	};
+};
