+++ src/server/model/project.js
@@ -1,0 +1,11 @@
+'use strict';
+
+import { projectQueryBuilder } from '../../db/helpers/projectQueryBuilder';
+
+module.exports = knex => {
+	const models = projectQueryBuilder(knex);
+
+	return {
+		...models
+	};
+};
